check directories

+++ initial-cp3-deploy.sh
@@ -51,7 +51,11 @@ function setUpNginx {
     sudo locale-gen en_US.UTF-8
 
     #remove the nginx default pages
-    sudo rm /etc/nginx/sites-enabled/default /etc/nginx/sites-available/default
+    #check if directory exists then delete
+    if [[ -d /etc/nginx/sites-enabled/default ]] && [[ -d /etc/nginx/sites-available/default ]] ; then
+        echo " nginx default directory exists. Deleting it..."
+        sudo rm /etc/nginx/sites-enabled/default /etc/nginx/sites-available/default
+    fi
 
     #create a proxy config file and a symlink to it in sites enabled
     sudo touch /etc/nginx/sites-available/brightevents.com
